requirements.txt
python-dotenv
langchain-core
langchain-community
openai
rank-bm25
ragas

